erster schritt für value bitstream implementiert

+++ JPEG/JPEG/Test.cs
@@ -639,8 +639,8 @@ namespace JPEG
         {
             byte[] arry = {57,45,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,23,0,0,30,1,0,0,0,0,0 };
             RunLengthEncoder coder = new RunLengthEncoder(arry);
-            byte[] test = coder.encodeACRunLength();
-            foreach (byte value in test)
+            short[] test = coder.encodeACRunLength();
+            foreach (short value in test)
             {
                 System.Console.WriteLine(value);
             }
